[Refactor] Request/Response 디렉토리 분리
