+++ keewe-api/src/main/java/ccc/keeweapi/component/CommentAssembler.java
@@ -1,10 +1,10 @@
 package ccc.keeweapi.component;
 
-import ccc.keeweapi.dto.insight.CommentResponse;
-import ccc.keeweapi.dto.insight.CommentWriterResponse;
-import ccc.keeweapi.dto.insight.InsightCommentCountResponse;
-import ccc.keeweapi.dto.insight.PreviewCommentResponse;
-import ccc.keeweapi.dto.insight.ReplyResponse;
+import ccc.keeweapi.dto.insight.response.CommentResponse;
+import ccc.keeweapi.dto.insight.response.CommentWriterResponse;
+import ccc.keeweapi.dto.insight.response.InsightCommentCountResponse;
+import ccc.keeweapi.dto.insight.response.PreviewCommentResponse;
+import ccc.keeweapi.dto.insight.response.ReplyResponse;
 import ccc.keewedomain.persistence.domain.insight.Comment;
 import ccc.keewedomain.persistence.domain.user.User;
 import org.springframework.stereotype.Component;
