+++ keewe-api/src/main/java/ccc/keeweapi/dto/insight/response/CommentResponse.java
@@ -1,17 +1,23 @@
-package ccc.keeweapi.dto.insight;
+package ccc.keeweapi.dto.insight.response;
 
 import ccc.keeweapi.dto.BlockFilteringResponse;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
+import java.util.List;
+
+
 @Getter
 @AllArgsConstructor(staticName = "of")
-public class PreviewCommentResponse implements BlockFilteringResponse {
+public class CommentResponse implements BlockFilteringResponse {
+
     private Long id;
     private CommentWriterResponse writer;
     private String content;
     private String createdAt;
+    private List<ReplyResponse> replies;
+    private Long totalReply;
 
     @Override
     @JsonIgnore
