+++ keewe-api/src/main/java/ccc/keeweapi/dto/insight/response/CommentWriterResponse.java
@@ -1,4 +1,4 @@
-package ccc.keeweapi.dto.insight;
+package ccc.keeweapi.dto.insight.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
