+++ keewe-api/src/main/java/ccc/keeweapi/dto/insight/response/InsightViewIncrementResponse.java
@@ -1,4 +1,4 @@
-package ccc.keeweapi.dto.insight;
+package ccc.keeweapi.dto.insight.response;
 
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
