+++ keewe-api/src/main/java/ccc/keeweapi/dto/insight/response/ReactResponse.java
@@ -1,4 +1,4 @@
-package ccc.keeweapi.dto.insight;
+package ccc.keeweapi.dto.insight.response;
 
 import ccc.keewedomain.persistence.domain.insight.enums.ReactionType;
 import lombok.AccessLevel;
