+++ keewe-api/src/main/java/ccc/keeweapi/dto/insight/response/ReplyResponse.java
@@ -1,4 +1,4 @@
-package ccc.keeweapi.dto.insight;
+package ccc.keeweapi.dto.insight.response;
 
 import ccc.keeweapi.dto.BlockFilteringResponse;
 import com.fasterxml.jackson.annotation.JsonIgnore;
