+++ keewe-api/src/main/java/ccc/keeweapi/service/insight/query/InsightDrawerQueryApiService.java
@@ -1,7 +1,7 @@
 package ccc.keeweapi.service.insight.query;
 
 import ccc.keeweapi.component.InsightAssembler;
-import ccc.keeweapi.dto.insight.DrawerResponse;
+import ccc.keeweapi.dto.insight.request.DrawerResponse;
 import ccc.keeweapi.utils.BlockFilterUtil;
 import ccc.keeweapi.utils.SecurityUtil;
 import ccc.keewedomain.service.insight.DrawerDomainService;
